requirements.txt
PyYAML>=6.0
Jinja2>=3.1.2
WeasyPrint>=60.0
pydantic>=2.0.0
python-dateutil
questionary>=2.0.0